refactor(query): move the 1730-line test module into executor/tests.rs
mod.rs 3712 -> 1986 lines. Pure code move: mod.rs keeps `#[cfg(test)] mod tests;`
and the body lives in tests.rs (still a child of executor, so use super::* and
private access resolve unchanged). Architecture-roadmap item [M/BR4], slice 2.
